Created Project and Updated Python
I was failing to create a scrapy project and this was due to using an old version of scrapy that was not playing well with my installed versions of python namely, 3.10 and 3.11. I recreated the virtual environment, installed the latest version of scrapy and that remedied the issue. I forgot to add a gitignore file and ended up pushing the virtual environment to source control so this is an attempt to remedy that problem.